Extract Message to a component and allow pin/unpin
